Moved duplicate methods to BlockCipherSupport class.
