+++ PNGAnalyzer/PNGAnalyzer/PNGAnalyzer/BlockCiphers/CipherBlockChaining.cs
@@ -27,19 +27,19 @@ namespace PNGAnalyzer.BlockCiphers
 
         public List<Chunk> CipherImage(List<Chunk> chunks)
         {
-            byte[] decompressedBytes = DecompressIDATs(chunks);
+            byte[] decompressedBytes = BlockCipherSupport.DecompressIDATs(chunks);
             byte[] cipheredBytes = Cipher(decompressedBytes);
-            List<Chunk> resultIdats = CompressIDATs(cipheredBytes);
-            List<Chunk> resultChunks = SwapIDATs(chunks, resultIdats);
+            List<Chunk> resultIdats = BlockCipherSupport.CompressIDATs(cipheredBytes);
+            List<Chunk> resultChunks = BlockCipherSupport.SwapIDATs(chunks, resultIdats);
             return resultChunks;
         }
 
         public byte[] Cipher(byte[] data)
         {
-            byte[] dataToDivide = AddPadding(data);
-            List<byte[]> blocks = DivideIntoBlocks(dataToDivide, BlockSize);
+            byte[] dataToDivide = BlockCipherSupport.AddPadding(data, BlockSize);
+            List<byte[]> blocks = BlockCipherSupport.DivideIntoBlocks(dataToDivide, BlockSize);
             CipherBlocks(blocks);
-            return ConcatenateBlocks(blocks);
+            return BlockCipherSupport.ConcatenateBlocks(blocks);
         }
 
         private void CipherBlocks(List<byte[]> blocks)
@@ -53,75 +53,6 @@ namespace PNGAnalyzer.BlockCiphers
                 blocks[i] = BigIntegerExtensions.UnsignedToBytes(block);
                 blocks[i] = rsa.Encrypt(blocks[i]);
             }
-        }
-
-        private static byte[] AddPadding(byte[] data)
-        {
-            int remainder = data.Length % BlockSize;
-            int bytesToAdd = remainder != 0 ? BlockSize - remainder : BlockSize;
-            byte[] dataWithPadding = AppendPadding(data, bytesToAdd);
-            return dataWithPadding;
-        }
-
-        private static byte[] AppendPadding(byte[] data, int bytesToAdd)
-        {
-            byte[] dataWithPadding = new byte[data.Length + bytesToAdd];
-            byte[] dataToAppend = new byte[bytesToAdd];
-            dataToAppend[dataToAppend.Length - 1] = Convert.ToByte(bytesToAdd);
-            data.CopyTo(dataWithPadding, 0);
-            dataToAppend.CopyTo(dataWithPadding, data.Length);
-            return dataWithPadding;
-        }
-
-        private static List<byte[]> DivideIntoBlocks(byte[] dataToDivide, int blockSize)
-        {
-            List<byte[]> result = new List<byte[]>();
-            for (int i = 0; i < dataToDivide.Length; i += blockSize)
-                result.Add(dataToDivide.Skip(i).Take(blockSize).ToArray());
-
-            return result;
-        }
-
-        private byte[] ConcatenateBlocks(List<byte[]> blocks)
-        {
-            int dataLength = blocks.Sum(block => block.Length);
-            byte[] data = new byte[dataLength];
-            int index = 0;
-            foreach (byte[] block in blocks)
-            {
-                block.CopyTo(data, index);
-                index += block.Length;
-            }
-
-            return data;
-        }
-
-        private static byte[] DecompressIDATs(List<Chunk> chunks)
-        {
-            List<IDAT> idats = chunks.Where(IsIDAT).Select(chunk => (IDAT) chunk).ToList();
-            byte[] bytes = IDATConverter.ConcatToBytes(idats);
-            byte[] decompressedBytes = ZlibCompression.Decompress(bytes);
-            return decompressedBytes;
-        }
-
-        private static bool IsIDAT(Chunk chunk)
-        {
-            return chunk.Type == "IDAT";
-        }
-
-        private static List<Chunk> CompressIDATs(byte[] cipheredBytes)
-        {
-            byte[] compressedBytes = ZlibCompression.Compress(cipheredBytes);
-            List<Chunk> resultIdats = IDATConverter.SplitToIDATs(compressedBytes).Select(idat => (Chunk) idat).ToList();
-            return resultIdats;
-        }
-
-        private static List<Chunk> SwapIDATs(List<Chunk> chunks, List<Chunk> resultIdats)
-        {
-            int firstIdatIndex = chunks.TakeWhile(chunk => !IsIDAT(chunk)).Count();
-            List<Chunk> resultChunks = chunks.Where(chunk => !IsIDAT(chunk)).ToList();
-            resultChunks.InsertRange(firstIdatIndex, resultIdats);
-            return resultChunks;
         }
     }
 }
